# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для твоего проекта",
        "title": "Job Assistant API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/telegram": {
            "post": {
                "description": "Логин через Telegram WebApp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Автоизация"
                ],
                "summary": "Telegram авторизация",
                "parameters": [
                    {
                        "description": "Telegram init data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.AuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.AuthResponse"
                        }
                    }
                }
            }
        },
        "/me": {
            "get": {
                "description": "Возвращает текущего пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Текущий пользователь",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.MeResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.AuthRequest": {
            "type": "object",
            "properties": {
                "init_data": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "language_code": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "telegram_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "views.AuthResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
        "views.MeResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        }
    }
}